Less opinionated about constructor of abstract class
